requirements.txt
flask==2.0.2
python-binance==1.0.16
python-dotenv==0.19.2
pandas==1.3.5
numpy==1.21.5
apscheduler==3.9.1
